requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
Pillow>=9.0.0
opencv-python>=4.5.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.5

# StyleGAN3 Requirements
click>=8.0
requests>=2.26.0
tqdm>=4.62.0
ninja>=1.10.2
psutil>=5.8.0

# PSPNet Requirements
dlib>=19.22.0
lpips>=0.1.4
gdown>=4.4.0

# Face Processing
face-alignment>=1.3.5
scikit-image>=0.19.0

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Utilities
pandas>=1.3.0
pyyaml>=6.0
joblib>=1.1.0

# Development (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0

# Documentation (optional)
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0